requirements.txt
ultralytics==8.2.103
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.2.0
yt-dlp>=2024.8.6
deep-sort-realtime>=1.3.2
moviepy>=1.0.3
PyYAML>=6.0.1
tqdm>=4.66.4
shapely